Set up render loop logic, working on spin
Still working on spin right now. The main problem I'm having is that the drag gesture doesn't actually change the velocity of the bouncer, only the position, strictly speaking. So I cannot read the instantaneous velocity to apply torque to the ball and instead have to pseudo-apply a velocity to the bouncer in the handlePanning method.