Layout2, Sacrifice Life
Added in a sacrifice life button to reset the ball's velocity and position when it becomes stuck moving in one dimension or plane. Found two bugs that I wasn't able to solve, and both of them are a result of spin: when the ball is spinning and it hits something, sometimes it registers it as more than one collision. This becomes a problem when it comes to the score because it can addScore()'s more than once per brick. Same with when the ball hits the bound. The life count can drop by more than one by just one fall